Create TestingGlueHiveMetastore
