+++ plugin/trino-delta-lake/src/test/java/io/trino/plugin/deltalake/TestDeltaLakeSharedGlueMetastoreViews.java
@@ -17,7 +17,7 @@ import io.trino.plugin.hive.metastore.HiveMetastore;
 
 import java.nio.file.Path;
 
-import static io.trino.plugin.hive.metastore.glue.GlueHiveMetastore.createTestingGlueHiveMetastore;
+import static io.trino.plugin.hive.metastore.glue.TestingGlueHiveMetastore.createTestingGlueHiveMetastore;
 
 /**
  * Requires AWS credentials, which can be provided any way supported by the DefaultProviderChain
